% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artlm.con.R
\name{artlm.con}
\alias{artlm.con}
\title{Per-Term Linear Model on Data Aligned-and-Ranked with ART-C}
\usage{
artlm.con(m, f, response = "art", factor.contrasts = "contr.sum", ...)
}
\arguments{
\item{m}{An object of class \code{\link{art}}.}

\item{f}{An object of type \code{"character"} indicating the effect term in
the transformed data in \code{m} to use as the aligned or art response.
this the same as the \code{"term"} in \code{\link{artlm}}}

\item{response}{Which response to use: the aligned (with ART-C) response
(\code{"aligned"}) or the aligned and ranked (with ART-C) response
(\code{"art"}).}

\item{factor.contrasts}{The name of the contrast-generating function to be
applied by default to fixed effect factors. Sets the the first element of
\code{\link{options}("contrasts")} for the duration of this function. The
default is to use \code{"contr.sum"}, i.e. sum-to-zero contrasts, which is
appropriate for Type III ANOVAs (the default ANOVA type for
\code{\link{anova.art}}).}

\item{\dots}{Additional arguments passed to \code{\link{lm}} or
\code{\link{lmer}}.}
}
\value{
An object of class \code{\link{lm}} if \code{formula(m)} does not
  contain grouping or error terms, an object of class \code{\link{merMod}}
  (i.e. a model fit by \code{\link{lmer}}) if it contains grouping terms, or
  an object of class \code{aovlist} (i.e. a model fit by \code{\link{aov}})
  if it contains error terms.
}
\description{
Given an \code{\link{art}} model, build a linear model from data aligned or
aligned-and-ranked with ART-C alignment procedure by the specified term in
the model.
}
\details{
This function is used primarily to conduct contrasts on data
aligned-and-ranked using the ART-C procedure using a different contrast test
than provided by \code{\link{artcon}}, which uses (\code{\link{contrast}}).
It is not necessary to use this function directly to conduct contrasts using
the ART-C procedure.

Internally, the ART-C procedure concatenates the variables specified
  in \code{f}, and then removes the originals. When specifying the effect
  terms on which to conduct contrasts, use the concatenation of the effects
  specified in \code{f} instead of the original variables. This is demonstrated
  in the example below.
}
\examples{
\dontrun{

## create an art model
m <- art(DryMatter ~ Moisture*Fertilizer + (1|Tray), data=Higgins1990Table5)

## use emmeans to conduct pariwise contrasts on "Moisture"
contrast(emmeans(artlm.con(m, "Moisture"), pairwise ~ Moisture))

## use emmeans to conduct pairwise contrasts on "Moisture:Fertilizer"
## N.B. internally, artlm.con concatenates the factors Moisture and Fertilizer
## to create MoistureFertilizer. If you try to use any of Moisture, Fertilizer,
## Moisture:Fertilizer, or Moisture*Fertilizer in the RHS of the formula
## passed to emmeans, you will get an error because the factors Moisture and Fertilizer
## do not exist in the model returned by artlm.con.
contrast(emmeans(artlm.con(m, "Moisture:Fertilizer"), pairwise ~ MoistureFertilizer))

## Note: artcon uses emmeans internally and the above examples are equivalent to
artcon(m, "Moisture")
artcon(m, "Moisture:Fertilizer")

}

}
\references{
Lisa A. Elkin, Matthew Kay, James J. Higgins, Jacob O. Wobbrock.
Under review (CHI 2021).
}
\seealso{
See also \code{\link{artcon}}, which makes use of this function.
}
\author{
Lisa A. Elkin
}
