% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Higgins1990Table1.art}
\alias{Higgins1990Table1.art}
\title{Aligned Rank Transformed Version of Higgins1990Table1}
\format{A data frame with 36 observations on the following 10 variables.
\describe{ \item{Subject}{a numeric vector} \item{Row}{a
numeric vector} \item{Column}{a numeric vector}
\item{Response}{a numeric vector}
\item{aligned.Response..for.Row}{a numeric vector}
\item{aligned.Response..for.Column}{a numeric vector}
\item{aligned.Response..for.Row.Column}{a numeric vector}
\item{ART.Response..for.Row}{a numeric vector}
\item{ART.Response..for.Column}{a numeric vector}
\item{ART.Response..for.Row.Column}{a numeric vector} }}
\source{
Wobbrock, J. O., Findlater, L., Gergle, D., and Higgins, J. J.
\emph{ARTool}. \url{http://depts.washington.edu/aimgroup/proj/art/}.
}
\description{
The ART version of \code{\link{Higgins1990Table1}} as produced by the
original ARTool, used to test the correctness of \code{\link{art}} output.
}
\seealso{
\code{\link{Higgins1990Table1}}, \code{\link{art}}.
}
\keyword{datasets}
\keyword{internal}

