% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HigginsABC.art}
\alias{HigginsABC.art}
\title{Aligned Rank Transformed Version of HigginsABC}
\format{A data frame with 16 observations on the following 19 variables.
\describe{ \item{Subject}{a numeric vector} \item{A}{a
numeric vector} \item{B}{a numeric vector} \item{C}{a
numeric vector} \item{Y}{a numeric vector}
\item{aligned.Y..for.A}{a numeric vector}
\item{aligned.Y..for.B}{a numeric vector}
\item{aligned.Y..for.A.B}{a numeric vector}
\item{aligned.Y..for.C}{a numeric vector}
\item{aligned.Y..for.A.C}{a numeric vector}
\item{aligned.Y..for.B.C}{a numeric vector}
\item{aligned.Y..for.A.B.C}{a numeric vector}
\item{ART.Y..for.A}{a numeric vector} \item{ART.Y..for.B}{a
numeric vector} \item{ART.Y..for.A.B}{a numeric vector}
\item{ART.Y..for.C}{a numeric vector}
\item{ART.Y..for.A.C}{a numeric vector}
\item{ART.Y..for.B.C}{a numeric vector}
\item{ART.Y..for.A.B.C}{a numeric vector} }}
\source{
Wobbrock, J. O., Findlater, L., Gergle, D., and Higgins, J. J.
\emph{ARTool}. \url{http://depts.washington.edu/aimgroup/proj/art/}.
}
\description{
The ART version of \code{\link{HigginsABC}} as produced by the original
ARTool, used to test the correctness of \code{\link{art}} output.
}
\seealso{
\code{\link{HigginsABC}}, \code{\link{art}}.
}
\keyword{datasets}
\keyword{internal}

