\name{InteractionTestData}
\alias{InteractionTestData}
\docType{data}
\title{
Synthetic Data Used in the Contrast Test Vignette
}
\description{
See (\code{vignette("art-contrasts")} for a description of this data.
}
\usage{data(InteractionTestData)}
\seealso{
\code{\link{art}}, \code{\link{anova.art}}. 
}
\examples{\dontrun{
## see this vignette
vignette("art-contrasts")
}}
\keyword{datasets}
