% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted-app.R
\name{restricted_app}
\alias{restricted_app}
\title{R Shiny application to simulate clinical trials with restricted randomization}
\usage{
restricted_app()
}
\description{
R Shiny application to simulate clinical trials with restricted randomization
}
