% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted.R
\name{rr}
\alias{rr}
\title{Restricted randomization}
\usage{
rr(nsbj, w, proc, proc_param, distr, distr_param)
}
\arguments{
\item{nsbj}{number of subjects to randomize.}

\item{w}{vector of fixed allocation ratio (integers with GCD = 1).}

\item{proc}{name of randomization procedure \code{c("CRD", "PBD", "BUD", "MWUD", 
"DBCD", "DL", "MinQD", "MaxEnt")}.}

\item{proc_param}{tuning parameter of randomization procedure (NA for \code{"CRD"}).}

\item{distr}{distribution of response \code{c("binary", "uniform", "normal", 
"exponential", "weibull", "loglogistic", "lognormal")}. See \link[artool]{response}
for details.}

\item{distr_param}{parameter(s) of the response distribution (a list with numeric vector(s) with the
same length as \code{w}). See \link[artool]{response}.}
}
\value{
a list with three items: 
   \itemize{
      \item \code{treatement} -- a vector of treatment assignments,
      \item \code{response} -- a vector of responses,
      \item \code{Imb} -- a vector of imbalances at each step,
      \item \code{FI} -- a vector of forcing indecies at each step,
      \item \code{MPM1} -- a evvctor of momentums of probability mass based on expected
            \emph{predicted} imbalance,
      \item \code{MPM2} -- a vector of momentums of probability mass based on the expected
             imbalance,
      \item \code{probability} -- a matrix with allocation probabilities at each step,
      \item \code{allocation} -- a matrix of allocation proportions at each step.
   }
}
\description{
Simulates a clinical trial one time with retricted randomizaion procedure
}
\examples{
   rr(200, c(4, 3, 2, 1), "DL", 2, "normal", list(mean = rep(0, 4), sd = rep(1, 4)))
   rr(200, c(31, 13), "MaxEnt", 0.5, "normal", list(mean = c(0, 0.2), sd = c(1, 1))) 
      
}
\author{
Yevgen Ryeznik (\email{yevgen.ryeznik@gmail.com}), Oleksandr Sverdlov
}
