% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example1.R
\name{run_example1}
\alias{run_example1}
\title{simulation example for a paper in "Statistics in Medicine"}
\usage{
run_example1(nsim)
}
\arguments{
\item{nsim}{number of simulations}
}
\description{
simulation example for a paper in "Statistics in Medicine"
}
