% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted.R
\name{simulate_rr}
\alias{simulate_rr}
\title{Restricted randomization}
\usage{
simulate_rr(nsim, nsbj, w, proc, proc_param, distr, distr_param)
}
\arguments{
\item{nsim}{number of simulation runs.}

\item{nsbj}{number of subjects to randomize.}

\item{w}{vector of fixed allocation ratio (integers with GCD = 1).}

\item{proc}{name of randomization procedure \code{c("CRD", "PBD", "BUD", "MWUD", 
"DBCD", "DL", "MinQD", "MaxEnt")}.}

\item{proc_param}{tuning parameter of randomization procedure (NA for \code{"CRD"}).}

\item{distr}{distribution of response \code{c("binary", "uniform", "normal", 
"exponential", "weibull", "loglogistic", "lognormal")}. See \link[artool]{response}
for details.}

\item{distr_param}{parameter(s) of the response distribution (a list with numeric vector(s) with the
same length as \code{w}). See \link[artool]{response}.}
}
\value{
a list with three items: 
   \itemize{
      \item \code{treatment} -- a matrix of treatment assignments with \code{nsim} rows
      and \code{nsbj} columns, i.e. r-th row contains tretament assignments of r-th 
      simulation,
      \item \code{response} -- a matrix of responses with \code{nsim} rows
      and \code{nsbj} columns, i.e. r-th row contains responses of r-th 
      simulation, 
      \item{\code{op} -- a data frame which contains operating characteristics:
          \itemize{
             \item{\code{MI} -- maximum imbalance,}
             \item{\code{AFI} -- average forcing index,}
             \item{\code{AMPM1} -- average momentum of probability mass based on expected \emph{predicted} imbalance,}
             \item{\code{ACMPM1} -- average cumulative momentum of probability mass based on expected \emph{predicted} imbalance,}
             \item{\code{AMPM2} -- average momentum of probability mass based on the expected imbalance,}
             \item{\code{ACMPM2} -- average cumulative momentum of probability mass based on expected \emph{predicted} imbalance,}
             \item{\code{ASD} -- average standard deviation of allocation proportions,}
         }
      }
      \item{\code{probability} -- a data frame with unconditional allocation probabilities vs number of subjects
   for each treatment;}
      \item{\code{allocation} -- a data frame with allocation proportions in the end of a trial
   vs number of simulations for each treatment (could be usefull to make a boxplot of allocation proportion distribution).}
   }
}
\description{
Makes multiple simulates of a clinical trial with retricted randomizaion procedures
}
\examples{
   simulate_rr(1000, 200, c(4, 3, 2, 1), "DL", 2, "normal", list(mean = rep(0, 4), sd = rep(1, 4)))
   simulate_rr(1000, 200, c(31, 13), "MaxEnt", 0.5, "normal", list(mean = c(0, 0.2), sd = c(1, 1)))    
}
\author{
Yevgen Ryeznik (\email{yevgen.ryeznik@gmail.com}), Oleksandr Sverdlov
}
