\name{Higgins1990Table1.art}
\alias{Higgins1990Table1.art}
\docType{data}
\title{
Aligned Rank Transformed Version of Higgins1990Table1
}
\description{
The ART version of \code{\link{Higgins1990Table1}} as produced by the original ARTool, used to test
the correctness of \code{\link{art}} output.
}
\usage{data(Higgins1990Table1.art)}
\format{
  A data frame with 36 observations on the following 10 variables.
  \describe{
    \item{\code{Subject}}{a numeric vector}
    \item{\code{Row}}{a numeric vector}
    \item{\code{Column}}{a numeric vector}
    \item{\code{Response}}{a numeric vector}
    \item{\code{aligned.Response..for.Row}}{a numeric vector}
    \item{\code{aligned.Response..for.Column}}{a numeric vector}
    \item{\code{aligned.Response..for.Row.Column}}{a numeric vector}
    \item{\code{ART.Response..for.Row}}{a numeric vector}
    \item{\code{ART.Response..for.Column}}{a numeric vector}
    \item{\code{ART.Response..for.Row.Column}}{a numeric vector}
  }
}
\source{
Wobbrock, J. O., Findlater, L., Gergle, D., and Higgins, J. J. \emph{ARTool}. \url{http://depts.washington.edu/aimgroup/proj/art/}.
}
\seealso{
\code{\link{Higgins1990Table1}}, \code{\link{art}}.
}
\keyword{datasets}
\keyword{internal}
