\name{Higgins1990Table5.art}
\alias{Higgins1990Table5.art}
\docType{data}
\title{
Aligned Rank Transformed Version of Higgins1990Table5
}
\description{
The ART version of \code{\link{Higgins1990Table5}} as produced by the original ARTool, used to test
the correctness of \code{\link{art}} output.
}
\usage{data(Higgins1990Table5.art)}
\format{
  A data frame with 48 observations on the following 10 variables.
  \describe{
    \item{\code{Tray}}{a numeric vector}
    \item{\code{Moisture}}{a numeric vector}
    \item{\code{Fertilizer}}{a numeric vector}
    \item{\code{DryMatter}}{a numeric vector}
    \item{\code{aligned.DryMatter..for.Moisture}}{a numeric vector}
    \item{\code{aligned.DryMatter..for.Fertilizer}}{a numeric vector}
    \item{\code{aligned.DryMatter..for.Moisture.Fertilizer}}{a numeric vector}
    \item{\code{ART.DryMatter..for.Moisture}}{a numeric vector}
    \item{\code{ART.DryMatter..for.Fertilizer}}{a numeric vector}
    \item{\code{ART.DryMatter..for.Moisture.Fertilizer}}{a numeric vector}
  }
}
\source{
Wobbrock, J. O., Findlater, L., Gergle, D., and Higgins, J. J. \emph{ARTool}. \url{http://depts.washington.edu/aimgroup/proj/art/}.
}
\seealso{
\code{\link{Higgins1990Table5}}, \code{\link{art}}.
}
\keyword{datasets}
\keyword{internal}
