% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{InteractionTestData}
\alias{InteractionTestData}
\title{Synthetic Data Used in the Contrast Test Vignette}
\description{
See (\code{vignette("art-contrasts")} for a description of this data.
}
\examples{
\dontrun{
## see this vignette
vignette("art-contrasts")
}
}
\seealso{
\code{\link{art}}, \code{\link{anova.art}}.
}
\keyword{datasets}
